+++ src/App.jsx
@@ -4,8 +4,6 @@ import SkillsSection from "./components/SkillsSection";
 import WorkSection from "./components/WorkSection";
 import WorkList from "./data/WorkList";
 import AboutSection from "./components/AboutSection";
-import SkillsSectionSmall from "./components/SkillsSectionSmall";
-import SkillsSectionSmall2 from "./components/SkillsSectionSmall2";
 
 function App() {
   return (
@@ -17,8 +15,6 @@ function App() {
             <AboutSection />
             <SkillsSection />
             <WorkSection src={WorkList} />
-            <SkillsSectionSmall />
-            <SkillsSectionSmall2 />
             <Footer />
           </main>
         </div>
